Refactor: API 클라이언트 패키지 구조 및 메서드명 수정
- global.client
-> global.client.clova
-> global.client.openai

- 프로퍼티도 client와 같은 패키지에서 관리하기 위함
